+++ pyemma/coordinates/estimation/covariance.py
@@ -21,6 +21,7 @@ import numpy as np
 import numbers
 from math import log
 
+from pyemma._base.serialization.serialization import SerializableMixIn
 from pyemma.util.annotators import deprecated
 from pyemma.util.types import is_float_vector, ensure_float_vector
 from pyemma.coordinates.data._base.streaming_estimator import StreamingEstimator
@@ -33,7 +34,10 @@ __all__ = ['LaggedCovariance']
 __author__ = 'paul, nueske'
 
 
-class LaggedCovariance(StreamingEstimator):
+class LaggedCovariance(StreamingEstimator, SerializableMixIn):
+    __serialize_version = 0
+    __serialize_fields = []
+
     r"""Compute lagged covariances between time series.
 
      Parameters
@@ -105,7 +109,6 @@ class LaggedCovariance(StreamingEstimator):
                         weights=weights, stride=stride, skip=skip, ncov_max=ncov_max)
 
         self._rc = None
-        self._used_data = 0
 
     @property
     def weights(self):
@@ -147,7 +150,7 @@ class LaggedCovariance(StreamingEstimator):
                 self.logger.info("adapting storage size")
                 self.nsave = nsave
         else: # in case we do a one shot estimation, we want to re-initialize running_covar
-            self._logger.debug("using %s moments for %i chunks", nsave, n_chunks)
+            self.logger.debug("using %s moments for %i chunks", nsave, n_chunks)
             self._rc = running_covar(xx=self.c00, xy=self.c0t, yy=self.ctt,
                                      remove_mean=self.remove_data_mean, symmetrize=self.reversible,
                                      sparse_mode=self.sparse_mode, modify_data=self.modify_data, nsave=nsave)
@@ -223,7 +226,11 @@ class LaggedCovariance(StreamingEstimator):
                 pg.update(1, stage=0)
 
         if partial_fit:
-            self._used_data += len(it)
+            if '_rc' not in self.__serialize_fields:
+                self.__serialize_fields.append('_rc')
+        else:
+            if '_rc' in self.__serialize_fields:
+                self.__serialize_fields.remove('_rc')
 
     def partial_fit(self, X):
         """ incrementally update the estimates