+++ pyemma/coordinates/tests/test_pipeline.py
@@ -80,17 +80,6 @@ class TestPipeline(unittest.TestCase):
         p.parametrize()
         self.assertTrue(p._is_estimated(), "If parametrized was called, the pipeline should be parametrized.")
 
-    def test_notify_changes_mixin(self):
-        X_t = np.random.random((30,30))
-        source = coor.source(np.array(X_t))
-
-        t1 = coor.tica(source)
-        from pyemma.coordinates.transform import TICA
-        t2 = TICA(lag=10)
-        assert len(t1._stream_children) == 0
-        t2.data_producer = t1
-        assert t1._stream_children[0] == t2
-
     def test_np_reader_in_pipeline(self):
         with TemporaryDirectory() as td:
             file_name = os.path.join(td, "test.npy")