+++ pyemma/coordinates/transform/tica.py
@@ -249,7 +249,6 @@ class TICA(StreamingEstimationTransformer, SerializableMixIn):
                                         cov=self._covar.C00_,
                                         cov_tau=self._covar.C0t_)
 
-        self._used_data = self._covar._used_data
         self._estimated = False
 
         return self
@@ -272,7 +271,6 @@ class TICA(StreamingEstimationTransformer, SerializableMixIn):
         self._model.update_model_params(mean=covar.mean,
                                         cov=covar.C00_,
                                         cov_tau=covar.C0t_)
-        self._used_data = covar._used_data
         self._diagonalize()
 
         return self._model
@@ -297,7 +295,7 @@ class TICA(StreamingEstimationTransformer, SerializableMixIn):
 
     def _diagonalize(self):
         # diagonalize with low rank approximation
-        self._logger.debug("diagonalize Cov and Cov_tau.")
+        self.logger.debug("diagonalize Cov and Cov_tau.")
         try:
             eigenvalues, eigenvectors = eig_corr(self.cov, self.cov_tau, self.epsilon, sign_maxelement=True)
         except ZeroRankError:
@@ -312,7 +310,7 @@ class TICA(StreamingEstimationTransformer, SerializableMixIn):
             regularized_timescales = 0.5 * timescales * np.maximum(np.tanh(np.pi * ((timescales - self.lag) / self.lag) + 1), 0)
 
             eigenvectors *= np.sqrt(regularized_timescales / 2)
-        self._logger.debug("finished diagonalisation.")
+        self.logger.debug("finished diagonalisation.")
 
         # compute cumulative variance
         cumvar = np.cumsum(np.abs(eigenvalues) ** 2)